Add option to include receiver ids into aircraft.json
This commit adds two new command line switches:
--aircraft-json-seen-by-list
--aircraft-json-seen-by-list-timeout

If --aircraft-json-seen-by-list is given, the receiver id of feeders
that have provided data for an aircraft will be included in the
corresponding aircraft object in aircraft.json along with the last timestamp.
The --aircraft-json-seen-by-list-timeout controls how many seconds after
the last update a receiver id will be included.